+++ packages/backend-output-storage/src/store_attribution_metadata.ts
@@ -1,7 +1,6 @@
 import { Stack } from 'aws-cdk-lib';
 import * as _os from 'os';
-import { CDKContextKey } from '@aws-amplify/platform-core';
-import * as _fs from 'fs';
+import { CDKContextKey, PackageJsonReader } from '@aws-amplify/platform-core';
 import { DeploymentType } from '@aws-amplify/plugin-types';
 
 /**
@@ -12,8 +11,8 @@ export class AttributionMetadataStorage {
    * Constructor with props for injecting test mocks
    */
   constructor(
-    private readonly fs: typeof _fs = _fs,
-    private readonly os: typeof _os = _os
+    private readonly os: typeof _os = _os,
+    private readonly packageJsonReader = new PackageJsonReader()
   ) {}
 
   /**
@@ -51,29 +50,11 @@ export class AttributionMetadataStorage {
   ): AttributionMetadata => ({
     createdOn: this.getPlatform(),
     createdBy: this.getDeploymentEngineType(stack),
-    createdWith: this.getLibraryVersion(libraryPackageJsonAbsolutePath),
+    createdWith:
+      this.packageJsonReader.read(libraryPackageJsonAbsolutePath).version ?? '', // This shouldn't happen, regardless we shouldn't throw for attribution
     stackType: stackType,
     metadata: additionalMetadata,
   });
-
-  private getLibraryVersion = (absolutePackageJsonPath: string): string => {
-    if (!this.fs.existsSync(absolutePackageJsonPath)) {
-      throw new Error(
-        `Could not find ${absolutePackageJsonPath} to load library version from`
-      );
-    }
-    const packageJsonContents = JSON.parse(
-      // we have to use sync fs methods here because this is part of cdk synth
-      this.fs.readFileSync(absolutePackageJsonPath, 'utf-8')
-    );
-    const libraryVersion = packageJsonContents.version;
-    if (typeof libraryVersion !== 'string') {
-      throw new Error(
-        `Could not parse library version from ${absolutePackageJsonPath}`
-      );
-    }
-    return libraryVersion;
-  };
 
   private getDeploymentEngineType = (stack: Stack): DeploymentEngineType => {
     const deploymentType: DeploymentType | undefined = stack.node.tryGetContext(