+++ packages/client-config/src/client-config-types/mobile/client_config_mobile_types.ts
@@ -43,6 +43,15 @@ export type ClientConfigMobileAuth = {
       Auth: {
         Default: {
           authenticationFlowType: 'USER_SRP_AUTH';
+          mfaConfiguration: string | undefined;
+          mfaTypes: Array<string> | undefined;
+          passwordProtectionSettings: {
+            passwordPolicyMinLength: number | undefined;
+            passwordPolicyCharacters: Array<string>;
+          };
+          signupAttributes: Array<string>;
+          usernameAttributes: Array<string>;
+          verificationMechanisms: Array<string>;
         };
       };
       AppSync?: {