+++ packages/client-config/src/client-config-writer/client_config_converter.test.ts
@@ -18,6 +18,25 @@ void describe('client config converter', () => {
       aws_cognito_region: 'test_cognito_region',
       aws_user_pools_id: 'test_user_pool_id',
       aws_user_pools_web_client_id: 'test_user_pool_app_client_id',
+      aws_cognito_identity_pool_id: 'test_identity_pool_id',
+      aws_cognito_signup_attributes: [
+        'test_signup_attribute_1',
+        'test_signup_attribute_2',
+      ],
+      aws_cognito_username_attributes: [
+        'test_username_attribute_1',
+        'test_username_attribute_2',
+      ],
+      aws_cognito_password_protection_settings: {
+        passwordPolicyMinLength: 1234,
+        passwordPolicyCharacters: ['a', 'b', 'c'],
+      },
+      aws_cognito_verification_mechanisms: [
+        'test_verification_mechanism_1',
+        'test_verification_mechanism_2',
+      ],
+      aws_cognito_mfa_configuration: 'test_mfa_configuration',
+      aws_cognito_mfa_types: ['test_mfa_type_1', 'test_mfa_type_2'],
     };
     const expectedMobileConfig: ClientConfigMobile = {
       UserAgent: expectedUserAgent,
@@ -37,7 +56,7 @@ void describe('client config converter', () => {
             CredentialsProvider: {
               CognitoIdentity: {
                 Default: {
-                  PoolId: 'test_user_pool_id',
+                  PoolId: 'test_identity_pool_id',
                   Region: 'test_cognito_region',
                 },
               },
@@ -45,6 +64,24 @@ void describe('client config converter', () => {
             Auth: {
               Default: {
                 authenticationFlowType: 'USER_SRP_AUTH',
+                mfaConfiguration: 'test_mfa_configuration',
+                mfaTypes: ['test_mfa_type_1', 'test_mfa_type_2'],
+                signupAttributes: [
+                  'test_signup_attribute_1',
+                  'test_signup_attribute_2',
+                ],
+                usernameAttributes: [
+                  'test_username_attribute_1',
+                  'test_username_attribute_2',
+                ],
+                passwordProtectionSettings: {
+                  passwordPolicyMinLength: 1234,
+                  passwordPolicyCharacters: ['a', 'b', 'c'],
+                },
+                verificationMechanisms: [
+                  'test_verification_mechanism_1',
+                  'test_verification_mechanism_2',
+                ],
               },
             },
           },
@@ -91,6 +128,7 @@ void describe('client config converter', () => {
     const clientConfig: ClientConfig = {
       aws_cognito_region: 'test_cognito_region',
       aws_user_pools_id: 'test_user_pool_id',
+      aws_cognito_identity_pool_id: 'test_identity_pool_id',
       aws_user_pools_web_client_id: 'test_user_pool_app_client_id',
       aws_appsync_region: 'test_app_sync_region',
       aws_appsync_graphqlEndpoint: 'https://test_api_endpoint.amazon.com',
@@ -115,7 +153,7 @@ void describe('client config converter', () => {
             CredentialsProvider: {
               CognitoIdentity: {
                 Default: {
-                  PoolId: 'test_user_pool_id',
+                  PoolId: 'test_identity_pool_id',
                   Region: 'test_cognito_region',
                 },
               },
@@ -123,6 +161,15 @@ void describe('client config converter', () => {
             Auth: {
               Default: {
                 authenticationFlowType: 'USER_SRP_AUTH',
+                mfaConfiguration: undefined,
+                mfaTypes: undefined,
+                signupAttributes: [],
+                usernameAttributes: [],
+                passwordProtectionSettings: {
+                  passwordPolicyCharacters: [],
+                  passwordPolicyMinLength: undefined,
+                },
+                verificationMechanisms: [],
               },
             },
             AppSync: {