+++ packages/integration-tests/test-projects/data-iterative-deploy/update-1/data/resource.ts
@@ -1,0 +1,10 @@
+import { defineData } from '@aws-amplify/backend';
+
+export const data = defineData({
+  schema: /* GraphQL */ `
+    type Todo @model {
+      field1: String! @index
+      field2: String! @index
+    }
+  `,
+});