+++ packages/backend-deployer/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @aws-amplify/backend-deployer
 
+## 0.3.1
+
+### Patch Changes
+
+- 70685f36b: Add usage data metrics
+- Updated dependencies [70685f36b]
+  - @aws-amplify/platform-core@0.2.1
+
 ## 0.3.0
 
 ### Minor Changes